WorkflowFunction class refactor PR-1: result-contract extraction.
